requirements.txt
bidict==0.22.0
blinker==1.4
cffi==1.15.0
click==8.1.3
colorama==0.4.4
Deprecated==1.2.13
dnspython==1.16.0
eventlet==0.30.2
Flask==2.1.2
flask-buzz==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-praetorian==1.3.0
flask-sock==0.5.2
Flask-SocketIO==5.1.2
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
importlib-metadata==4.11.3
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pendulum==2.1.2
psycopg2==2.9.3
py-buzz==2.1.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytzdata==2020.1
simple-websocket==0.5.2
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
